% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-reactive.R
\name{dropXreactiveLoci}
\alias{dropXreactiveLoci}
\title{Drop cross-reactive probes/loci}
\usage{
dropXreactiveLoci(x)
}
\arguments{
\item{x}{A \code{minfi} object}
}
\value{
An object of the same kind as the input, possibly with fewer
loci.
}
\description{
Remove cross-reactive probes from
}
\examples{
\dontrun{
suppressPackageStartupMessages(library(minfiData))
data(MsetEx)
MsetEx_noXloci <- dropXreactiveLoci(MsetEx)
nrow(MsetEx) - nrow(MsetEx_noXloci)
}
}
