% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_array_type}
\alias{get_array_type}
\title{Get array type}
\usage{
get_array_type(x)
}
\arguments{
\item{x}{An object containing annotation information}
}
\value{
A character scale of either "EPIC" or "450K"
}
\description{
Given an object containing annotation information, the function gets
array type which is used by other internal functions.
}
