% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-reactive.R
\name{read_xreactive_probes_450k_benton}
\alias{read_xreactive_probes_450k_benton}
\title{Read cross-reactive probes of 450K array from Benton et. al.}
\usage{
read_xreactive_probes_450k_benton()
}
\value{
A character vector of probe IDs
}
\description{
The cross-reactive probes are from the paper \href{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-014-0569-x}{Benton et al. (2015)}.
}
\details{
The cross-reactive probes are downloaded from file
\code{HumanMethylation450_15017482_v.1.1_hg19_bowtie_multimap.txt} of
\href{https://github.com/sirselim/illumina450k_filtering}{GitHub: sirselim/illumina450k_filtering}.
}
