% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-reactive.R
\name{read_xreactive_probes_450k_chen}
\alias{read_xreactive_probes_450k_chen}
\title{Read cross-reactive probes of 450K array from Chen et. al.}
\usage{
read_xreactive_probes_450k_chen()
}
\value{
A character vector of probe IDs
}
\description{
The cross-reactive probes are from the paper \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3592906/}{Chen et al. (2013)}.
}
\details{
The cross-reactive probes are downloaded from file
\code{48639-non-specific-probes-Illumina450k.csv} of
\href{https://github.com/sirselim/illumina450k_filtering}{GitHub: sirselim/illumina450k_filtering}.
}
