% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-reactive.R
\name{read_xreactive_probes_epic_mccartney}
\alias{read_xreactive_probes_epic_mccartney}
\title{Read cross-reactive probes of 850K/EPIC array from McCartney et. al.}
\usage{
read_xreactive_probes_epic_mccartney()
}
\value{
A character vector of probe IDs
}
\description{
The cross-reactive probes are originally published in paper
\href{https://www.sciencedirect.com/science/article/pii/S221359601630071X}{McCartney et al. (2016)}.
}
\details{
The cross-reactive probes are downloaded from Table S2 and S3.
}
