% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-reactive.R
\name{read_xreactive_probes_epic_pidsley}
\alias{read_xreactive_probes_epic_pidsley}
\title{Read cross-reactive probes of 850K/EPIC array from Pidsley et. al.}
\usage{
read_xreactive_probes_epic_pidsley()
}
\value{
A character vector of probe IDs
}
\description{
The cross-reactive probes are originally published in paper
\href{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-016-1066-1}{Pidsley et al. (2016)}.
}
\details{
The cross-reactive probes are downloaded from file
\code{EPIC/13059_2016_1066_MOESM1_ESM.csv} of
\href{https://github.com/sirselim/illumina450k_filtering}{GitHub: sirselim/illumina450k_filtering}.
}
