% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-reactive.R
\name{xreactive_probes}
\alias{xreactive_probes}
\title{Cross-reactive probes/loci}
\usage{
xreactive_probes(array_type = "EPIC")
}
\arguments{
\item{array_type}{A character scalar. One of "EPIC", "450K". Defaults
to "EPIC".}
}
\value{
A character vector of probe IDs
}
\description{
Get cross-reactive probe/locus IDs from internal data
}
\details{
The cross-reactive probes/loci of EPIC/850K are from papers:

\enumerate{
  \item \href{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-016-1066-1}{Pidsley et al., Critical evaluation of the Illumina MethylationEPIC BeadChip microarray for whole-genome DNA methylation profiling. Genome Biology (2016)}
  \item \href{https://www.sciencedirect.com/science/article/pii/S221359601630071X}{McCartney et al., Identification of polymorphic and off-target probe binding sites on the Illumina Infinium MethylationEPIC BeadChip. Genomics Data (2016)}
}

The cross-reactive probes/loci of 450K are from papers:

\enumerate{
  \item \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3592906/}{Chen et al., Discovery of cross-reactive probes and polymorphic CpGs in the Illumina Infinium HumanMethylation450 microarray. Epigenetics (2013)}
  \item \href{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-014-0569-x}{Benton et al., An analysis of DNA methylation in human adipose tissue reveals differential modification of obesity genes before and after gastric bypass and weight loss. Genome Biology (2015)}
}
}
\examples{
xloci <- xreactive_probes()
}
